- Fehler bei der Duplikaterkennungbehoben - Duplikaterkennungverbessert - Daten werden jetzt tatsächlich gespeichert und geladen - Konformitätsprüfung der URL's verbessert - Dokumentation verbessert

+++ app/src/main/java/firesoft/de/kalenderadapter/data/ServerParameter.java
@@ -81,7 +81,7 @@ public class ServerParameter implements AbstractMap.Entry {
             this.value = String.valueOf(input);
         }
         else if (value instanceof Integer) {
-            this.value = (Integer) input;
+            this.value = input;
         }
 
         return null;
